# packages/backend/src/controllers/runQueryController.ts
import {
    AdditionalMetric,
    ApiErrorPayload,
    ApiQueryResults,
    CacheMetadata,
    CustomDimension,
    DateGranularity,
    Explore,
    FieldId,
    Item,
    MetricQuery,
    MetricQueryRequest,
    MetricQueryResponse,
    SortField,
    TableCalculation,
} from '@lightdash/common';
import {
    Body,
    Controller,
    Middlewares,
    OperationId,
    Path,
    Post,
    Request,
    Response,
    Route,
    SuccessResponse,
    Tags,
} from '@tsoa/runtime';
import express from 'express';
import { projectService } from '../services/services';
import { allowApiKeyAuthentication, isAuthenticated } from './authentication';

export type ApiRunQueryResponse = {
    status: 'ok';
    results: {
        metricQuery: MetricQueryResponse; // tsoa doesn't support complex types like MetricQuery
        cacheMetadata: CacheMetadata;
        rows: any[];
        fields?: Record<string, Item | AdditionalMetric>;
    };
};

@Route('/api/v1/projects/{projectUuid}')
@Response<ApiErrorPayload>('default', 'Error')
@Tags('Exploring')
export class RunViewChartQueryController extends Controller {
    /**
     * Run a query for underlying data results
     * @param projectUuid The uuid of the project
     * @param body metricQuery for the chart to run
     * @param exploreId table name
     * @param req express request
     */
    @Middlewares([allowApiKeyAuthentication, isAuthenticated])
    @SuccessResponse('200', 'Success')
    @Post('/explores/{exploreId}/runUnderlyingDataQuery')
    @OperationId('postRunUnderlyingDataQuery')
    async postUnderlyingData(
        @Body() body: MetricQueryRequest,
        @Path() projectUuid: string,
        @Path() exploreId: string,
        @Request() req: express.Request,
    ): Promise<ApiRunQueryResponse> {
        const metricQuery: MetricQuery = {
            exploreName: body.exploreName,
            dimensions: body.dimensions,
            metrics: body.metrics,
            filters: body.filters,
            sorts: body.sorts,
            limit: body.limit,
            tableCalculations: body.tableCalculations,
            additionalMetrics: body.additionalMetrics,
            customDimensions: body.customDimensions,
        };
        const results: ApiQueryResults =
            await projectService.runUnderlyingDataQuery(
                req.user!,
                metricQuery,
                projectUuid,
                exploreId,
                body.csvLimit,
            );
        this.setStatus(200);
        return {
            status: 'ok',
            results,
        };
    }

    /**
     * Run a query for explore
     * @param projectUuid The uuid of the project
     * @param body metricQuery for the chart to run
     * @param exploreId table name
     * @param req express request
     */
    @Middlewares([allowApiKeyAuthentication, isAuthenticated])
    @SuccessResponse('200', 'Success')
    @Post('/explores/{exploreId}/runQuery')
    @OperationId('RunMetricQuery')
    async runMetricQuery(
        @Body() body: MetricQueryRequest,
        @Path() projectUuid: string,
        @Path() exploreId: string,

        @Request() req: express.Request,
    ): Promise<ApiRunQueryResponse> {
        const metricQuery: MetricQuery = {
            exploreName: body.exploreName,
            dimensions: body.dimensions,
            metrics: body.metrics,
            filters: body.filters,
            sorts: body.sorts,
            limit: body.limit,
            tableCalculations: body.tableCalculations,
            additionalMetrics: body.additionalMetrics,
            customDimensions: body.customDimensions,
        };
        const results: ApiQueryResults = await projectService.runExploreQuery(
            req.user!,
            metricQuery,
            projectUuid,
            exploreId,
            body.csvLimit,
            body.granularity,
        );
        this.setStatus(200);
        return {
            status: 'ok',
            results,
        };
    }

    /**
     * Run a query for explore
     * @param projectUuid The uuid of the project
     * @param body metricQuery for the chart to run
     * @param exploreId table name
     * @param req express request
     */
    @Middlewares([allowApiKeyAuthentication, isAuthenticated])
    @SuccessResponse('200', 'Success')
    @Post('/explores/runCustomExploreQuery')
    @OperationId('RunCustomExploreQuery')
    async runCustomExploreQuery(
        @Body()
        body: // TODO: improve types
        {
            metricQuery: any;
            explore: any;
            csvLimit?: number | undefined;
            granularity?: DateGranularity | undefined;
        },
        @Path() projectUuid: string,

        @Request() req: express.Request,
    ): Promise<ApiRunQueryResponse> {
        const results: ApiQueryResults =
            await projectService.runCustomExploreQuery(
                req.user!,
                projectUuid,
                req.body.metricQuery,
                req.body.explore,
                body.csvLimit,
                body.granularity,
            );
        this.setStatus(200);
        return {
            status: 'ok',
            results,
        };
    }
}