# packages/backend/src/generated/swagger.json
{
	"components": {
...
		"schemas": {
			"ApiErrorPayload": {
				"properties": {
					"error": {
						"properties": {
							"data": {
								"description": "Optional data containing details of the error"
							},
							"message": {
								"type": "string",
								"description": "A friendly message summarising the error"
							},
							"name": {
								"type": "string",
								"description": "Unique name for the type of error"
							},
							"statusCode": {
								"type": "number",
								"format": "integer",
								"description": "HTTP status code"
							}
						},
						"required": [
							"name",
							"statusCode"
						],
						"type": "object"
					},
					"status": {
						"type": "string",
						"enum": [
							"error"
						],
						"nullable": false
					}
				},
				"required": [
					"error",
					"status"
				],
				"type": "object",
				"description": "The Error object is returned from the api any time there is an error.\nThe message contains"
			},
...
			"FieldId": {
				"type": "string"
			},
...
			"ConditionalOperator": {
				"enum": [
					"isNull",
					"notNull",
					"equals",
					"notEquals",
					"startsWith",
					"endsWith",
					"include",
					"doesNotInclude",
					"lessThan",
					"lessThanOrEqual",
					"greaterThan",
					"greaterThanOrEqual",
					"inThePast",
					"notInThePast",
					"inTheNext",
					"inTheCurrent",
					"inBetween"
				],
				"type": "string"
			},
...
			"SortField": {
				"properties": {
					"descending": {
						"type": "boolean"
					},
					"fieldId": {
						"type": "string"
					}
				},
				"required": [
					"descending",
					"fieldId"
				],
				"type": "object"
			},
			"CustomFormatType": {
				"enum": [
					"default",
					"percent",
					"currency",
					"number",
					"id"
				],
				"type": "string"
			},
			"NumberSeparator": {
				"enum": [
					"default",
					"commaPeriod",
					"spacePeriod",
					"periodComma",
					"noSeparatorPeriod"
				],
				"type": "string"
			},
			"Compact": {
				"enum": [
					"thousands",
					"millions",
					"billions",
					"trillions"
				],
				"type": "string"
			},
			"CompactOrAlias": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/Compact"
					},
					{
						"type": "string",
						"enum": [
							"K",
							"thousand",
							"M",
							"million",
							"B",
							"billion",
							"T",
							"trillion"
						]
					}
				]
			},
			"CustomFormat": {
				"properties": {
					"type": {
						"$ref": "#/components/schemas/CustomFormatType"
					},
					"round": {
						"type": "number",
						"format": "double"
					},
					"separator": {
						"$ref": "#/components/schemas/NumberSeparator"
					},
					"currency": {
						"type": "string"
					},
					"compact": {
						"$ref": "#/components/schemas/CompactOrAlias"
					},
					"prefix": {
						"type": "string"
					},
					"suffix": {
						"type": "string"
					}
				},
				"required": [
					"type"
				],
				"type": "object",
				"additionalProperties": true
			},
			"TableCalculation": {
				"properties": {
					"format": {
						"$ref": "#/components/schemas/CustomFormat"
					},
					"sql": {
						"type": "string"
					},
					"displayName": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"index": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"sql",
					"displayName",
					"name"
				],
				"type": "object"
			},
			"MetricType": {
				"enum": [
					"percentile",
					"average",
					"count",
					"count_distinct",
					"sum",
					"min",
					"max",
					"number",
					"median",
					"string",
					"date",
					"timestamp",
					"boolean"
				],
				"type": "string"
			},
			"Format": {
				"enum": [
					"km",
					"mi",
					"usd",
					"gbp",
					"eur",
					"id",
					"percent"
				],
				"type": "string"
			},
			"MetricFilterRule": {
				"properties": {
					"values": {
						"items": {},
						"type": "array"
					},
					"operator": {
						"$ref": "#/components/schemas/ConditionalOperator"
					},
					"id": {
						"type": "string"
					},
					"target": {
						"properties": {
							"fieldRef": {
								"type": "string"
							}
						},
						"required": [
							"fieldRef"
						],
						"type": "object"
					},
					"settings": {},
					"disabled": {
						"type": "boolean"
					}
				},
				"required": [
					"operator",
					"id",
					"target"
				],
				"type": "object",
				"additionalProperties": true
			},
			"AdditionalMetric": {
				"properties": {
					"label": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/MetricType"
					},
					"description": {
						"type": "string"
					},
					"sql": {
						"type": "string"
					},
					"hidden": {
						"type": "boolean"
					},
					"round": {
						"type": "number",
						"format": "double"
					},
					"compact": {
						"$ref": "#/components/schemas/CompactOrAlias"
					},
					"format": {
						"$ref": "#/components/schemas/Format"
					},
					"table": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"index": {
						"type": "number",
						"format": "double"
					},
					"filters": {
						"items": {
							"$ref": "#/components/schemas/MetricFilterRule"
						},
						"type": "array"
					},
					"baseDimensionName": {
						"type": "string"
					},
					"uuid": {
						"type": "string",
						"nullable": true
					},
					"percentile": {
						"type": "number",
						"format": "double"
					},
					"formatOptions": {
						"$ref": "#/components/schemas/CustomFormat"
					}
				},
				"required": [
					"type",
					"sql",
					"table",
					"name"
				],
				"type": "object",
				"additionalProperties": true
			},
			"BinType": {
				"enum": [
					"fixed_number",
					"fixed_width",
					"custom_range"
				],
				"type": "string"
			},
			"BinRange": {
				"properties": {
					"to": {
						"type": "number",
						"format": "double"
					},
					"from": {
						"type": "number",
						"format": "double"
					}
				},
				"type": "object"
			},
			"CustomDimension": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"dimensionId": {
						"$ref": "#/components/schemas/FieldId"
					},
					"table": {
						"type": "string"
					},
					"binType": {
						"$ref": "#/components/schemas/BinType"
					},
					"binNumber": {
						"type": "number",
						"format": "double"
					},
					"binWidth": {
						"type": "number",
						"format": "double"
					},
					"customRange": {
						"items": {
							"$ref": "#/components/schemas/BinRange"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"dimensionId",
					"table",
					"binType"
				],
				"type": "object",
				"additionalProperties": true
			},
			"Pick_CompiledDimension.label-or-name_": {
				"properties": {
					"name": {
						"type": "string"
					},
					"label": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"label"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
...
			"FilterGroupResponse": {
				"anyOf": [
					{
						"properties": {
							"or": {
								"items": {},
								"type": "array"
							},
							"id": {
								"type": "string"
							}
						},
						"required": [
							"or",
							"id"
						],
						"type": "object"
					},
					{
						"properties": {
							"and": {
								"items": {},
								"type": "array"
							},
							"id": {
								"type": "string"
							}
						},
						"required": [
							"and",
							"id"
						],
						"type": "object"
					}
				]
			},
			"FiltersResponse": {
				"properties": {
					"tableCalculations": {
						"$ref": "#/components/schemas/FilterGroupResponse"
					},
					"metrics": {
						"$ref": "#/components/schemas/FilterGroupResponse"
					},
					"dimensions": {
						"$ref": "#/components/schemas/FilterGroupResponse"
					}
				},
				"type": "object"
			},
			"MetricQueryResponse": {
				"properties": {
					"metadata": {
						"properties": {
							"hasADateDimension": {
								"$ref": "#/components/schemas/Pick_CompiledDimension.label-or-name_"
							}
						},
						"required": [
							"hasADateDimension"
						],
						"type": "object"
					},
					"customDimensions": {
						"items": {
							"$ref": "#/components/schemas/CustomDimension"
						},
						"type": "array"
					},
					"additionalMetrics": {
						"items": {
							"$ref": "#/components/schemas/AdditionalMetric"
						},
						"type": "array"
					},
					"tableCalculations": {
						"items": {
							"$ref": "#/components/schemas/TableCalculation"
						},
						"type": "array"
					},
					"limit": {
						"type": "number",
						"format": "double"
					},
					"sorts": {
						"items": {
							"$ref": "#/components/schemas/SortField"
						},
						"type": "array"
					},
					"filters": {
						"$ref": "#/components/schemas/FiltersResponse"
					},
					"metrics": {
						"items": {
							"$ref": "#/components/schemas/FieldId"
						},
						"type": "array"
					},
					"dimensions": {
						"items": {
							"$ref": "#/components/schemas/FieldId"
						},
						"type": "array"
					},
					"exploreName": {
						"type": "string"
					}
				},
				"required": [
					"tableCalculations",
					"limit",
					"sorts",
					"filters",
					"metrics",
					"dimensions",
					"exploreName"
				],
				"type": "object"
			},
...
			"ProjectType": {
				"enum": [
					"DEFAULT",
					"PREVIEW"
				],
				"type": "string"
			},
...
			"DbtProjectType.DBT": {
				"enum": [
					"dbt"
				],
				"type": "string"
			},
			"DbtProjectEnvironmentVariable": {
				"properties": {
					"value": {
						"type": "string"
					},
					"key": {
						"type": "string"
					}
				},
				"required": [
					"value",
					"key"
				],
				"type": "object"
			},
...
			"DbtLocalProjectConfig": {
				"properties": {
					"type": {
						"$ref": "#/components/schemas/DbtProjectType.DBT"
					},
					"target": {
						"type": "string"
					},
					"environment": {
						"items": {
							"$ref": "#/components/schemas/DbtProjectEnvironmentVariable"
						},
						"type": "array"
					},
					"profiles_dir": {
						"type": "string"
					},
					"project_dir": {
						"type": "string"
					}
				},
				"required": [
					"type"
				],
				"type": "object",
				"additionalProperties": true
			},
			"DbtProjectType.DBT_CLOUD_IDE": {
				"enum": [
					"dbt_cloud_ide"
				],
				"type": "string"
			},
			"DbtCloudIDEProjectConfig": {
				"properties": {
					"type": {
						"$ref": "#/components/schemas/DbtProjectType.DBT_CLOUD_IDE"
					},
					"api_key": {
						"type": "string"
					},
					"environment_id": {
						"type": "string"
					}
				},
				"required": [
					"type",
					"api_key",
					"environment_id"
				],
				"type": "object",
				"additionalProperties": true
			},
			"DbtProjectType.GITHUB": {
				"enum": [
					"github"
				],
				"type": "string"
			},
			"DbtGithubProjectConfig": {
				"properties": {
					"type": {
						"$ref": "#/components/schemas/DbtProjectType.GITHUB"
					},
					"target": {
						"type": "string"
					},
					"environment": {
						"items": {
							"$ref": "#/components/schemas/DbtProjectEnvironmentVariable"
						},
						"type": "array"
					},
					"personal_access_token": {
						"type": "string"
					},
					"repository": {
						"type": "string"
					},
					"branch": {
						"type": "string"
					},
					"project_sub_path": {
						"type": "string"
					},
					"host_domain": {
						"type": "string"
					}
				},
				"required": [
					"type",
					"personal_access_token",
					"repository",
					"branch",
					"project_sub_path"
				],
				"type": "object",
				"additionalProperties": true
			},
			"DbtProjectType.BITBUCKET": {
				"enum": [
					"bitbucket"
				],
				"type": "string"
			},
			"DbtBitBucketProjectConfig": {
				"properties": {
					"type": {
						"$ref": "#/components/schemas/DbtProjectType.BITBUCKET"
					},
					"target": {
						"type": "string"
					},
					"environment": {
						"items": {
							"$ref": "#/components/schemas/DbtProjectEnvironmentVariable"
						},
						"type": "array"
					},
					"username": {
						"type": "string"
					},
					"personal_access_token": {
						"type": "string"
					},
					"repository": {
						"type": "string"
					},
					"branch": {
						"type": "string"
					},
					"project_sub_path": {
						"type": "string"
					},
					"host_domain": {
						"type": "string"
					}
				},
				"required": [
					"type",
					"username",
					"personal_access_token",
					"repository",
					"branch",
					"project_sub_path"
				],
				"type": "object",
				"additionalProperties": true
			},
			"DbtProjectType.GITLAB": {
				"enum": [
					"gitlab"
				],
				"type": "string"
			},
			"DbtGitlabProjectConfig": {
				"properties": {
					"type": {
						"$ref": "#/components/schemas/DbtProjectType.GITLAB"
					},
					"target": {
						"type": "string"
					},
					"environment": {
						"items": {
							"$ref": "#/components/schemas/DbtProjectEnvironmentVariable"
						},
						"type": "array"
					},
					"personal_access_token": {
						"type": "string"
					},
					"repository": {
						"type": "string"
					},
					"branch": {
						"type": "string"
					},
					"project_sub_path": {
						"type": "string"
					},
					"host_domain": {
						"type": "string"
					}
				},
				"required": [
					"type",
					"personal_access_token",
					"repository",
					"branch",
					"project_sub_path"
				],
				"type": "object",
				"additionalProperties": true
			},
			"DbtProjectType.AZURE_DEVOPS": {
				"enum": [
					"azure_devops"
				],
				"type": "string"
			},
			"DbtAzureDevOpsProjectConfig": {
				"properties": {
					"type": {
						"$ref": "#/components/schemas/DbtProjectType.AZURE_DEVOPS"
					},
					"target": {
						"type": "string"
					},
					"environment": {
						"items": {
							"$ref": "#/components/schemas/DbtProjectEnvironmentVariable"
						},
						"type": "array"
					},
					"personal_access_token": {
						"type": "string"
					},
					"organization": {
						"type": "string"
					},
					"project": {
						"type": "string"
					},
					"repository": {
						"type": "string"
					},
					"branch": {
						"type": "string"
					},
					"project_sub_path": {
						"type": "string"
					}
				},
				"required": [
					"type",
					"personal_access_token",
					"organization",
					"project",
					"repository",
					"branch",
					"project_sub_path"
				],
				"type": "object",
				"additionalProperties": true
			},
			"DbtProjectType.NONE": {
				"enum": [
					"none"
				],
				"type": "string"
			},
			"DbtNoneProjectConfig": {
				"properties": {
					"type": {
						"$ref": "#/components/schemas/DbtProjectType.NONE"
					},
					"target": {
						"type": "string"
					},
					"environment": {
						"items": {
							"$ref": "#/components/schemas/DbtProjectEnvironmentVariable"
						},
						"type": "array"
					},
					"hideRefreshButton": {
						"type": "boolean"
					}
				},
				"required": [
					"type"
				],
				"type": "object",
				"additionalProperties": true
			},
			"DbtProjectConfig": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/DbtLocalProjectConfig"
					},
					{
						"$ref": "#/components/schemas/DbtCloudIDEProjectConfig"
					},
					{
						"$ref": "#/components/schemas/DbtGithubProjectConfig"
					},
					{
						"$ref": "#/components/schemas/DbtBitBucketProjectConfig"
					},
					{
						"$ref": "#/components/schemas/DbtGitlabProjectConfig"
					},
					{
						"$ref": "#/components/schemas/DbtAzureDevOpsProjectConfig"
					},
					{
						"$ref": "#/components/schemas/DbtNoneProjectConfig"
					}
				]
			},
			"WarehouseTypes.SNOWFLAKE": {
				"enum": [
					"snowflake"
				],
				"type": "string"
			},
			"WeekDay": {
				"enum": [
					0,
					1,
					2,
					3,
					4,
					5,
					6
				],
				"type": "number"
			},
			"Pick_CreateSnowflakeCredentials.Exclude_keyofCreateSnowflakeCredentials.SensitiveCredentialsFieldNames__": {
				"properties": {
					"warehouse": {
						"type": "string"
					},
					"type": {
						"$ref": "#/components/schemas/WarehouseTypes.SNOWFLAKE"
					},
					"role": {
						"type": "string"
					},
					"account": {
						"type": "string"
					},
					"requireUserCredentials": {
						"type": "boolean"
					},
					"database": {
						"type": "string"
					},
					"schema": {
						"type": "string"
					},
					"threads": {
						"type": "number",
						"format": "double"
					},
					"clientSessionKeepAlive": {
						"type": "boolean"
					},
					"queryTag": {
						"type": "string"
					},
					"accessUrl": {
						"type": "string"
					},
					"startOfWeek": {
						"allOf": [
							{
								"$ref": "#/components/schemas/WeekDay"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"warehouse",
					"type",
					"account",
					"database",
					"schema"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_CreateSnowflakeCredentials.SensitiveCredentialsFieldNames_": {
				"$ref": "#/components/schemas/Pick_CreateSnowflakeCredentials.Exclude_keyofCreateSnowflakeCredentials.SensitiveCredentialsFieldNames__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"SnowflakeCredentials": {
				"$ref": "#/components/schemas/Omit_CreateSnowflakeCredentials.SensitiveCredentialsFieldNames_"
			},
...
			"Omit_CreateRedshiftCredentials.SensitiveCredentialsFieldNames_": {
				"$ref": "#/components/schemas/Pick_CreateRedshiftCredentials.Exclude_keyofCreateRedshiftCredentials.SensitiveCredentialsFieldNames__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"RedshiftCredentials": {
				"$ref": "#/components/schemas/Omit_CreateRedshiftCredentials.SensitiveCredentialsFieldNames_"
			},
...
			"Omit_CreatePostgresCredentials.SensitiveCredentialsFieldNames_": {
				"$ref": "#/components/schemas/Pick_CreatePostgresCredentials.Exclude_keyofCreatePostgresCredentials.SensitiveCredentialsFieldNames__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"PostgresCredentials": {
				"$ref": "#/components/schemas/Omit_CreatePostgresCredentials.SensitiveCredentialsFieldNames_"
			},
...
			"Omit_CreateBigqueryCredentials.SensitiveCredentialsFieldNames_": {
				"$ref": "#/components/schemas/Pick_CreateBigqueryCredentials.Exclude_keyofCreateBigqueryCredentials.SensitiveCredentialsFieldNames__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"BigqueryCredentials": {
				"$ref": "#/components/schemas/Omit_CreateBigqueryCredentials.SensitiveCredentialsFieldNames_"
			},
...
			"Omit_CreateDatabricksCredentials.SensitiveCredentialsFieldNames_": {
				"$ref": "#/components/schemas/Pick_CreateDatabricksCredentials.Exclude_keyofCreateDatabricksCredentials.SensitiveCredentialsFieldNames__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"DatabricksCredentials": {
				"$ref": "#/components/schemas/Omit_CreateDatabricksCredentials.SensitiveCredentialsFieldNames_"
			},
...
			"Omit_CreateTrinoCredentials.SensitiveCredentialsFieldNames_": {
				"$ref": "#/components/schemas/Pick_CreateTrinoCredentials.Exclude_keyofCreateTrinoCredentials.SensitiveCredentialsFieldNames__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"TrinoCredentials": {
				"$ref": "#/components/schemas/Omit_CreateTrinoCredentials.SensitiveCredentialsFieldNames_"
			},
			"WarehouseCredentials": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/SnowflakeCredentials"
					},
					{
						"$ref": "#/components/schemas/RedshiftCredentials"
					},
					{
						"$ref": "#/components/schemas/PostgresCredentials"
					},
					{
						"$ref": "#/components/schemas/BigqueryCredentials"
					},
					{
						"$ref": "#/components/schemas/DatabricksCredentials"
					},
					{
						"$ref": "#/components/schemas/TrinoCredentials"
					}
				]
			},
			"SupportedDbtVersions": {
				"enum": [
					"v1.4",
					"v1.5",
					"v1.6",
					"v1.7"
				],
				"type": "string"
			},
			"Project": {
				"properties": {
					"dbtVersion": {
						"$ref": "#/components/schemas/SupportedDbtVersions"
					},
					"copiedFromProjectUuid": {
						"type": "string"
					},
					"pinnedListUuid": {
						"type": "string"
					},
					"warehouseConnection": {
						"$ref": "#/components/schemas/WarehouseCredentials"
					},
					"dbtConnection": {
						"$ref": "#/components/schemas/DbtProjectConfig"
					},
					"type": {
						"$ref": "#/components/schemas/ProjectType"
					},
					"name": {
						"type": "string"
					},
					"projectUuid": {
						"type": "string"
					},
					"organizationUuid": {
						"type": "string"
					}
				},
				"required": [
					"dbtVersion",
					"dbtConnection",
					"type",
					"name",
					"projectUuid",
					"organizationUuid"
				],
				"type": "object"
			},
			"ApiProjectResponse": {
				"properties": {
					"results": {
						"$ref": "#/components/schemas/Project"
					},
					"status": {
						"type": "string",
						"enum": [
							"ok"
						],
						"nullable": false
					}
				},
				"required": [
					"results",
					"status"
				],
				"type": "object"
			},
...
			"DateGranularity": {
				"enum": [
					"Day",
					"Week",
					"Month",
					"Quarter",
					"Year"
				],
				"type": "string"
			},
			"MetricQueryRequest": {
				"properties": {
					"metadata": {
						"properties": {
							"hasADateDimension": {
								"$ref": "#/components/schemas/Pick_CompiledDimension.label-or-name_"
							}
						},
						"required": [
							"hasADateDimension"
						],
						"type": "object"
					},
					"granularity": {
						"$ref": "#/components/schemas/DateGranularity"
					},
					"customDimensions": {
						"items": {
							"$ref": "#/components/schemas/CustomDimension"
						},
						"type": "array"
					},
					"csvLimit": {
						"type": "number",
						"format": "double"
					},
					"additionalMetrics": {
						"items": {
							"$ref": "#/components/schemas/AdditionalMetric"
						},
						"type": "array"
					},
					"tableCalculations": {
						"items": {
							"$ref": "#/components/schemas/TableCalculation"
						},
						"type": "array"
					},
					"limit": {
						"type": "number",
						"format": "double"
					},
					"sorts": {
						"items": {
							"$ref": "#/components/schemas/SortField"
						},
						"type": "array"
					},
					"filters": {
						"properties": {
							"tableCalculations": {},
							"metrics": {},
							"dimensions": {}
						},
						"type": "object"
					},
					"metrics": {
						"items": {
							"$ref": "#/components/schemas/FieldId"
						},
						"type": "array"
					},
					"dimensions": {
						"items": {
							"$ref": "#/components/schemas/FieldId"
						},
						"type": "array"
					},
					"exploreName": {
						"type": "string"
					}
				},
				"required": [
					"tableCalculations",
					"limit",
					"sorts",
					"filters",
					"metrics",
					"dimensions",
					"exploreName"
				],
				"type": "object"
			},
...
			"CacheMetadata": {
				"properties": {
					"cacheHit": {
						"type": "boolean"
					},
					"cacheUpdatedTime": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"cacheHit"
				],
				"type": "object"
			},
			"Record_string.Item-or-AdditionalMetric_": {
				"properties": {},
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
			"ApiRunQueryResponse": {
				"properties": {
					"results": {
						"properties": {
							"fields": {
								"$ref": "#/components/schemas/Record_string.Item-or-AdditionalMetric_"
							},
							"rows": {
								"items": {},
								"type": "array"
							},
							"cacheMetadata": {
								"$ref": "#/components/schemas/CacheMetadata"
							},
							"metricQuery": {
								"$ref": "#/components/schemas/MetricQueryResponse"
							}
						},
						"required": [
							"rows",
							"cacheMetadata",
							"metricQuery"
						],
						"type": "object"
					},
					"status": {
						"type": "string",
						"enum": [
							"ok"
						],
						"nullable": false
					}
				},
				"required": [
					"results",
					"status"
				],
				"type": "object"
			},
...
		},
		"securitySchemes": {
			"session_cookie": {
				"type": "apiKey",
				"in": "cookie",
				"name": "connect.sid"
			},
			"api_key": {
				"type": "apiKey",
				"in": "header",
				"name": "Authorization",
				"description": "Value should be 'ApiKey <your key>'"
			}
		}
	},
	"info": {
		"title": "Lightdash API",
		"version": "0.1021.4",
		"description": "Open API documentation for all public Lightdash API endpoints.  # Authentication Before you get started, you might need to create a Personal Access Token to authenticate via the API. You can create a token by following this guide: https://docs.lightdash.com/references/personal_tokens\n",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "Lightdash Support",
			"email": "support@lightdash.com",
			"url": "https://docs.lightdash.com/help-and-contact/contact/contact_info/"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/api/v1/projects/{projectUuid}": {
			"get": {
				"operationId": "GetProject",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiProjectResponse"
								}
							}
						}
					},
					"default": {
						"description": "Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorPayload"
								}
							}
						}
					}
				},
				"description": "Get a project of an organiztion",
				"tags": [
					"Projects"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "projectUuid",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/api/v1/projects/{projectUuid}/explores/{exploreId}/runUnderlyingDataQuery": {
			"post": {
				"operationId": "postRunUnderlyingDataQuery",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiRunQueryResponse"
								}
							}
						}
					},
					"default": {
						"description": "Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorPayload"
								}
							}
						}
					}
				},
				"description": "Run a query for underlying data results",
				"tags": [
					"Exploring"
				],
				"security": [],
				"parameters": [
					{
						"description": "The uuid of the project",
						"in": "path",
						"name": "projectUuid",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "table name",
						"in": "path",
						"name": "exploreId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "metricQuery for the chart to run",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/MetricQueryRequest",
								"description": "metricQuery for the chart to run"
							}
						}
					}
				}
			}
		},
		"/api/v1/projects/{projectUuid}/explores/{exploreId}/runQuery": {
			"post": {
				"operationId": "RunMetricQuery",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiRunQueryResponse"
								}
							}
						}
					},
					"default": {
						"description": "Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorPayload"
								}
							}
						}
					}
				},
				"description": "Run a query for explore",
				"tags": [
					"Exploring"
				],
				"security": [],
				"parameters": [
					{
						"description": "The uuid of the project",
						"in": "path",
						"name": "projectUuid",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "table name",
						"in": "path",
						"name": "exploreId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "metricQuery for the chart to run",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/MetricQueryRequest",
								"description": "metricQuery for the chart to run"
							}
						}
					}
				}
			}
		},
		"/api/v1/projects/{projectUuid}/explores/runCustomExploreQuery": {
			"post": {
				"operationId": "RunCustomExploreQuery",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiRunQueryResponse"
								}
							}
						}
					},
					"default": {
						"description": "Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiErrorPayload"
								}
							}
						}
					}
				},
				"description": "Run a query for explore",
				"tags": [
					"Exploring"
				],
				"security": [],
				"parameters": [
					{
						"description": "The uuid of the project",
						"in": "path",
						"name": "projectUuid",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "metricQuery for the chart to run",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"granularity": {
										"$ref": "#/components/schemas/DateGranularity"
									},
									"csvLimit": {
										"type": "number",
										"format": "double"
									},
									"explore": {},
									"metricQuery": {}
								},
								"required": [
									"explore",
									"metricQuery"
								],
								"type": "object",
								"description": "metricQuery for the chart to run"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	],
...
}